refactor(model): change uint to uint64 and int64

+++ internal/model/credit_card.go
@@ -3,10 +3,10 @@ package model
 import "time"
 
 type CreditCard struct {
-	ID             uint `gorm:"primaryKey"`
-	TelegramID     uint
-	Number         uint `gorm:"unique"`
+	ID             uint64 `gorm:"primaryKey"`
+	TelegramID     int64
+	Number         uint64 `gorm:"unique"`
 	Cardholder     string
 	ExpirationDate time.Time
-	CVV            uint
+	CVV            uint64
 }
