+++ internal/model/observer.go
@@ -1,8 +1,8 @@
 package model
 
 type Observer struct {
-	ID         uint `gorm:"primaryKey"`
-	TelegramID uint
-	FuelPumpID uint
+	ID         uint64 `gorm:"primaryKey"`
+	TelegramID int64
+	FuelPumpID uint64
 	GasStation FuelPump `gorm:"foreignKey:FuelPumpID"`
 }
